+++ queries/user/update_password.sql
@@ -1,6 +1,7 @@
 UPDATE
     user_account
 SET
-    hashed_password = $2
+    hashed_password = $2,
+    reset_token = NULL
 WHERE
     id = $1;