+++ src/site/patreon.rs
@@ -11,7 +11,7 @@ use uuid::Uuid;
 use crate::jobs::{JobContext, JobInitiatorExt};
 use crate::models::{LinkedAccount, Site};
 use crate::site::{CollectedSite, SiteFromConfig, SiteServices};
-use crate::{jobs, models, Error};
+use crate::{jobs, models, AsUrl, Error};
 
 pub struct Patreon {
     auth_url: AuthUrl,
@@ -209,6 +209,7 @@ async fn callback(
     config: web::Data<crate::Config>,
     conn: web::Data<sqlx::PgPool>,
     faktory: web::Data<FaktoryProducer>,
+    request: actix_web::HttpRequest,
     user: models::User,
     query: web::Query<types::PatreonOAuthCallback>,
 ) -> Result<HttpResponse, Error> {
@@ -398,7 +399,12 @@ async fn callback(
     }
 
     Ok(HttpResponse::Found()
-        .insert_header(("Location", format!("/user/account/{}", linked_account.id)))
+        .insert_header((
+            "Location",
+            request
+                .url_for("user_account", [linked_account.id.as_url()])?
+                .as_str(),
+        ))
         .finish())
 }